+++ tests/integration/test_access.py
@@ -1,7 +1,7 @@
 import json
 
 from click.testing import CliRunner
-from gen3_util.cli.cli import cli
+from gen3_util.repo.cli import cli
 
 
 def test_access_ls(caplog):
