+++ tests/integration/test_buckets.py
@@ -1,6 +1,6 @@
 
 from click.testing import CliRunner
-from gen3_util.cli.cli import cli
+from gen3_util.repo.cli import cli
 
 
 def test_project_ls(caplog):
